requirements.txt
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
pandas==1.2.3
drf-yasg==1.20.0